% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobContext.R
\name{CDJobContext}
\alias{CDJobContext}
\title{Subclass for implementation specific settings}
\description{
Subclass for implementation specific settings

Subclass for implementation specific settings
}
\section{Super class}{
\code{\link[CohortDiagnosticsModule:BaseJobContext]{CohortDiagnosticsModule::BaseJobContext}} -> \code{CDJobContext}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-validateSettings}{\code{CDJobContext$validateSettings()}}
\item \href{#method-getExecutionParameters}{\code{CDJobContext$getExecutionParameters()}}
\item \href{#method-clone}{\code{CDJobContext$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CohortDiagnosticsModule" data-topic="BaseJobContext" data-id="getCohortDefinitionSet">}\href{../../CohortDiagnosticsModule/html/BaseJobContext.html#method-getCohortDefinitionSet}{\code{CohortDiagnosticsModule::BaseJobContext$getCohortDefinitionSet()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CohortDiagnosticsModule" data-topic="BaseJobContext" data-id="getConnectionDetails">}\href{../../CohortDiagnosticsModule/html/BaseJobContext.html#method-getConnectionDetails}{\code{CohortDiagnosticsModule::BaseJobContext$getConnectionDetails()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CohortDiagnosticsModule" data-topic="BaseJobContext" data-id="initialize">}\href{../../CohortDiagnosticsModule/html/BaseJobContext.html#method-initialize}{\code{CohortDiagnosticsModule::BaseJobContext$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validateSettings"></a>}}
\if{latex}{\out{\hypertarget{method-validateSettings}{}}}
\subsection{Method \code{validateSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDJobContext$validateSettings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExecutionParameters"></a>}}
\if{latex}{\out{\hypertarget{method-getExecutionParameters}{}}}
\subsection{Method \code{getExecutionParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDJobContext$getExecutionParameters(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDJobContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
