% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnosticsModule.R
\docType{package}
\name{CohortDiagnosticsModule-package}
\alias{CohortDiagnosticsModule}
\alias{CohortDiagnosticsModule-package}
\title{CohortDiagnosticsModule: An R Package module for the OHDSI CohortDiagnostics package}
\description{
An R package for that provides a module for using the CohortGenerator package to generate cohorts against the OMOP CDM.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortDiagnosticsModule/}
  \item \url{https://github.com/OHDSI/CohortDiagnosticsModule}
  \item Report bugs at \url{https://github.com/OHDSI/CohortDiagnosticsModule/issues}
}

}
\author{
\strong{Maintainer}: Anthony Sena \email{sena@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
